requirements.txt
anyconfig
python-dotenv
netifaces